+++ gulpfile.js
@@ -35,7 +35,13 @@ var webpackConfig = {
     filename: LIB_JS,
     sourcePrefix: '  '
   },
+  module: {
+    loaders: [
+      {test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader'}
+    ]
+  },
   plugins: [ bannerPlugin ],
+  externals: {ws: 'ws'},
   cache: true
 };